refactor: providers of url
