+++ src/routes/Routes.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Route, Switch, Redirect } from 'react-router-dom';
-import BrowserURL from '../util/BrowserURL';
+import BrowserProvider from '../util/URL/BrowserProvider';
 
 import TodoList from '../pages/Todo/List/TodoList';
 import NotFound from '../pages/Fallback/NotFound';
@@ -10,11 +10,11 @@ import RouteValidator from '../auth/Routes/Validator';
 const Routes = () => {
   return (
     <Switch>
-      <Redirect path={BrowserURL.HOME} exact to={BrowserURL.TODO_LIST} />
-      <RouteValidator allowedAuthorities={['canAccessTodoListPage']} authorityKey="permissions" path={BrowserURL.TODO_LIST} component={TodoList} />
-      <Route path={BrowserURL.BUGGY} exact component={Buggy} />
-      <Route path={BrowserURL.NOT_FOUND} exact component={NotFound} />
-      <Route path={BrowserURL.ANY} component={NotFound} />
+      <Redirect path={BrowserProvider.HOME} exact to={BrowserProvider.TODO_LIST} />
+      <RouteValidator allowedAuthorities={['canAccessTodoListPage']} authorityKey="permissions" path={BrowserProvider.TODO_LIST} component={TodoList} />
+      <Route path={BrowserProvider.BUGGY} exact component={Buggy} />
+      <Route path={BrowserProvider.NOT_FOUND} exact component={NotFound} />
+      <Route path={BrowserProvider.ANY} component={NotFound} />
     </Switch>
   );
 };
