--- src/util/ApiURL.js
@@ -1,5 +1,0 @@
-export default {
-  GET_USER_LOGGED: '/v1/user/logged',
-  GET_TODO_BY_ID: '/v1/todo/detail/:id',
-  GET_TODO_LIST: '/v1/todo/all',
-};
